+++ packages/opentelemetry-instrumentation-llamaindex/opentelemetry/instrumentation/llamaindex/base_tool_instrumentor.py
@@ -4,6 +4,8 @@ from wrapt import wrap_function_wrapper
 
 from opentelemetry.instrumentation.llamaindex.utils import (
     _with_tracer_wrapper,
+    process_request,
+    process_response,
     start_as_current_span_async,
 )
 from opentelemetry.semconv.ai import SpanAttributes, TraceloopSpanKindValues
@@ -56,7 +58,10 @@ def query_wrapper(tracer, wrapped, instance, args, kwargs):
             TraceloopSpanKindValues.TOOL.value,
         )
 
-        return wrapped(*args, **kwargs)
+        process_request(span, args, kwargs)
+        res = wrapped(*args, **kwargs)
+        process_response(span, res)
+        return res
 
 
 @_with_tracer_wrapper
@@ -69,4 +74,7 @@ async def aquery_wrapper(tracer, wrapped, instance, args, kwargs):
             TraceloopSpanKindValues.TOOL.value,
         )
 
-        return await wrapped(*args, **kwargs)
+        process_request(span, args, kwargs)
+        res = await wrapped(*args, **kwargs)
+        process_response(span, res)
+        return res